fix(desktop): isolate window-parking FFI in a safe helper crate
The desktop crate forbids unsafe code, which the parking change violated
(NSWindow pointer derefs and a private-selector message). Move those three
calls into a workspace member with a minimal safe API that verifies the main
thread and obtains pointers from live Tauri windows; the app crate stays
free of unsafe and its forbid lint stands.
